+++ v4/src/core/config/AppConfig.js
@@ -7,7 +7,7 @@
   global.AppConfig = Object.freeze({
     app: Object.freeze({
       name: "My Produce Assistant",
-      version: "52.1.4"
+      version: "52.1.6"
     }),
 
     urls: Object.freeze({